absorbe the Pipeline into the dedicated Recognizer and Resolver base classes
