--- geoparser/pipeline/__init__.py
@@ -1,1 +1,0 @@
-from geoparser.pipeline.pipeline import Pipeline
